fix: correct target date in timer logic to 2025 and add InputClient component for self-reflection input
