+++ src/app/page.tsx
@@ -31,7 +31,7 @@ export default function Home() {
 
   useEffect(() => {
     const now = new Date();
-    const targetDate = new Date("2026-01-01T00:00:00+09:00"); // KST
+    const targetDate = new Date("2025-01-01T00:00:00+09:00"); // KST
     if (now >= targetDate) {
       setShowTimer(false); // 2026년이 지나면 타이머 숨기기
     }
